backend/python-usuarios-recomendaciones/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
pydantic-settings
beanie
